requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.4
psycopg2-binary==2.9.5
python-decouple==3.6
geoip2==2.9.0
pytz==2022.5
sqlparse==0.4.3
django-cors-headers==3.13.0
djangorestframework==3.14.0